Fixed percentage rounding

+++ console/data.go
@@ -93,7 +93,7 @@ func PrintCPUUsed(c *Console, n *Data, o *Data, s *vmguestlib.Session) {
 	if _, err := s.GetCPUUsed(); err != nil {
 		c.WriteNaCol()
 	} else {
-		c.WritePercentCol(float64((n.CPUUsed-o.CPUUsed)/(n.TimeElasped-o.TimeElasped)) * 100.0)
+		c.WritePercentCol(float64(n.CPUUsed-o.CPUUsed) / float64(n.TimeElasped-o.TimeElasped) * 100.0)
 	}
 }
 
@@ -101,7 +101,7 @@ func PrintCPUStolen(c *Console, n *Data, o *Data, s *vmguestlib.Session) {
 	if _, err := s.GetCPUStolen(); err != nil {
 		c.WriteNaCol()
 	} else {
-		c.WritePercentCol(float64((n.CPUStolen-o.CPUStolen)/(n.TimeElasped-o.TimeElasped)) * 100.0)
+		c.WritePercentCol(float64(n.CPUStolen-o.CPUStolen) / float64(n.TimeElasped-o.TimeElasped) * 100.0)
 	}
 }
 
@@ -137,7 +137,7 @@ func PrintHostCPUUsed(c *Console, n *Data, o *Data, s *vmguestlib.Session) {
 	if _, err := s.GetHostCPUUsed(); err != nil {
 		c.WriteNaCol()
 	} else {
-		c.WritePercentCol(float64((n.HostCPUUsed-o.HostCPUUsed)/(n.TimeElasped-o.TimeElasped)) * 100.0)
+		c.WritePercentCol(float64(n.HostCPUUsed-o.HostCPUUsed) / float64(n.TimeElasped-o.TimeElasped) * 100.0)
 	}
 }
